requirements.txt
Faker==18.3.1
kafka-python==2.0.2
boto3==1.40.38
streamlit==1.50.0
